requirements.txt
streamlit
openai
python-dotenv
pypdf
matplotlib
